MyCloudBackend/mycloud/wsgi.py
"""
WSGI config for mycloud project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/5.2/howto/deployment/wsgi/
"""

import os
from django.core.wsgi import get_wsgi_application

# ✅ ДОБАВЛЕНО: Загрузка переменных окружения для production
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    pass

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'mycloud.settings')

application = get_wsgi_application()
