MyCloudBackend/requirements.txt
asgiref==3.9.1
colorama==0.4.6
dj-database-url==2.1.0
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-extensions==3.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
factory-boy==3.3.0
Faker==37.8.0
gunicorn==21.2.0
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==7.4.2
pytest-django==4.5.2
pytest-mock==3.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
types-Pillow==10.2.0.20240822
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
